Fixing a RAM calculation issue

+++ stellarsolver/stellarsolver.h
@@ -269,7 +269,7 @@ class StellarSolver : public QObject
         SextractorSolver* createSextractorSolver();
 
         //This finds out the amount of available RAM on the system
-        uint64_t getAvailableRAM();
+        double getAvailableRAM();
         //This determines if there is enough RAM for the selected index files so that we don't try to load indexes inParallel unless it can handle it.
         bool enoughRAMisAvailableFor(QStringList indexFolders);
